Fixed the userId verification and other errors in the advert-controller

+++ applications/backend/src/controllers/advert-controller.js
@@ -81,14 +81,14 @@ export async function createAdvert(request, response) {
   } else if (MAX_DESCRIPTION_LENGTH < advertData.description.length) {
     errorMessages.push(
       "Description can't be more than " +
-        MAX_DESCRIPTION_LENGTH +
-        "characters long."
+      MAX_DESCRIPTION_LENGTH +
+      "characters long."
     );
   } else if (advertData.description.length < MIN_DESCRIPTION_LENGTH) {
     errorMessages.push(
       "Description can't be less than " +
-        MIN_DESCRIPTION_LENGTH +
-        " characters long."
+      MIN_DESCRIPTION_LENGTH +
+      " characters long."
     );
   }
 
@@ -167,7 +167,7 @@ export async function insertImageIntoAdvertById(request, response) {
   const updateID = advertID;
   try {
     const values = [image, updateID];
-    const insertedImageIntoAdvert = await db.query("UPDATE adverts SET img_src = ? WHERE advertID = ?", values);
+    await db.query("UPDATE adverts SET img_src = ? WHERE advertID = ?", values);
     response.status(200).send("Image insterted into advert successfully").json();
   } catch (error) {
     console.error(error.status);
@@ -181,43 +181,48 @@ export async function updateAdvertById(request, response) {
     response.status(400).send("Missing request body");
     return;
   }
+
   try {
     const authorizationHeaderValue = request.get("Authorization");
     const accessToken = authorizationHeaderValue.substring(7);
-    const isSigned = accessToken.split('.').length === 3;
+    const decodedToken = jwt.verify(accessToken, ACCESS_TOKEN_SECRET);
 
-    if (isSigned) {
-      const decodedToken = jwt.verify(accessToken, ACCESS_TOKEN_SECRET);
+    if (!decodedToken.isLoggedIn) {
+      throw new jwt.JsonWebTokenError();
+    }
 
-      if (!decodedToken.isLoggedIn) {
-        throw new jwt.JsonWebTokenError();
-      }
+    console.log("decodedToken.userId", decodedToken.userId)
+    console.log("advertData.accountID", advertData.accountID)
+    if (decodedToken.userId !== advertData.accountID) {
+      response.status(401).send("Unauthorized");
+      return;
     }
 
     const values = [advertData.title, advertData.description, advertData.price, request.params.id];
-    const updatedAccount = await db.query("UPDATE adverts SET title = ?, description = ?, price = ? WHERE advertID = ?", values);
-    response.status(200).send("Advert updated successfully").json();
+    await db.query("UPDATE adverts SET title = ?, description = ?, price = ? WHERE advertID = ?", values);
+    response.status(200).send("Advert updated successfully");
   } catch (error) {
-    console.error(error);
-    response.status(500).send("Internal server error");
+    if (error instanceof jwt.JsonWebTokenError) {
+      response.status(401).json([UNAUTHORIZED_USER_ERROR]);
+    } else {
+      console.error(error.status);
+      response.status(500).json([DATABASE_ERROR_MESSAGE]);
+    }
   }
 }
 
 export async function deleteAdvertById(request, response) {
+  console.log("deleteAdvertById")
   try {
     const authorizationHeaderValue = request.get("Authorization");
     const accessToken = authorizationHeaderValue.substring(7);
-    const isSigned = accessToken.split('.').length === 3;
+    const decodedToken = jwt.verify(accessToken, ACCESS_TOKEN_SECRET);
 
-    if (isSigned) {
-      const decodedToken = jwt.verify(accessToken, ACCESS_TOKEN_SECRET);
-
-      if (!decodedToken.isLoggedIn) {
-        throw new jwt.JsonWebTokenError();
-      }
+    if (!decodedToken.isLoggedIn) {
+      throw new jwt.JsonWebTokenError();
     }
 
-    await db.query("DELETE FROM adverts WHERE advertID = ?", [request.params.id])
+    await db.query("DELETE FROM adverts WHERE advertID = ? AND accountID = ?", [request.params.id, decodedToken.userId])
     response.status(204).send("Advert successfully deleted");
   } catch (error) {
     if (error instanceof jwt.JsonWebTokenError) {
